requirements.txt
discord
openai
python-decouple
chromadb
torch
scipy
peewee
munch
pynacl
transformers
phonemizer
ollama
diffusers
accelerate
txtai[pipeline]
git+https://github.com/neuml/txtai.git#egg=txtai